server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy<2.0
pymysql
python-dotenv
lxml
readability-lxml
requests
